+++ frontend/openchat-client/src/utils/chat.ts
@@ -87,7 +87,6 @@ import {
     currentUserIdStore,
     currentUserStore,
     localUpdates,
-    proposalTalliesStore,
     selectedChatIdStore,
     selectedChatUserIdsStore,
     selectedServerChatSummaryStore,
@@ -1375,9 +1374,7 @@ export function mergeEventsAndLocalUpdates(
 
             const tallyUpdate =
                 e.event.content.kind === "proposal_content"
-                    ? proposalTalliesStore.value.get(
-                          `${e.event.content.governanceCanisterId}_${e.event.content.proposal.id}`,
-                      )
+                    ? updates?.proposalTally
                     : undefined;
 
             const senderBlocked = selectedChatBlockedOrSuspendedUsers.has(e.event.sender);